requirements.txt
# =============================================================================
# DeFi Risk Lakehouse — Python dependencies
# Installed in both local dev and the custom Airflow Docker image
# =============================================================================

# HTTP / GraphQL
requests==2.32.3
gql[requests]==3.5.0

# AWS / MinIO
boto3==1.34.110
botocore==1.34.110

# Data processing
pyspark==3.5.1
pyarrow==16.1.0

# dbt
dbt-core==1.9.0
dbt-trino==1.9.0

# =============================================================================
# ML / MLOps dependencies (liquidation predictor extension)
# =============================================================================

# ML
xgboost==2.0.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# Experiment tracking + model registry
mlflow==2.13.0

# Hyperparameter optimization
optuna==3.6.1

# Interpretability
shap==0.45.1
matplotlib==3.9.0

# Model serving
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# Drift monitoring
evidently==0.4.30

# Data versioning
dvc==3.50.1
dvc-s3==3.2.0

# Postgres client (for predictions_log)
psycopg2-binary==2.9.9

# Trino Python client
trino==0.328.0
